+++ libopensimcreator/UI/ModelWarper/ModelWarperTab.cpp
@@ -1162,6 +1162,82 @@ Uses the Thin-Plate Spline (TPS) warping algorithm to scale `WrapCylinder`s in t
         }
     };
 
+    // A `ScalingStep` that scales body segments by the chosen (potentially, non-uniform) scale factors
+    class ManuallyScaleBodySegmentsScalingStep final : public ScalingStep {
+        OpenSim_DECLARE_CONCRETE_OBJECT(ManuallyScaleBodySegmentsScalingStep, ScalingStep)
+
+        OpenSim_DECLARE_LIST_PROPERTY(bodies, std::string, "Absolute paths (e.g. `/bodyset/femur`) to `Body` components in the model");
+        OpenSim_DECLARE_PROPERTY(scale_factors, SimTK::Vec3, "The manual scale factors that should be applied to the specified bodies");
+        OpenSim_DECLARE_PROPERTY(preserve_masses, bool, "If enabled, the masses of the bodies will not be scaled (inertial properties will still be scaled)");
+    public:
+        explicit ManuallyScaleBodySegmentsScalingStep() :
+            ScalingStep{"Manually Scale Body Segments"}
+        {
+            setDescription("Applies a manually-specified scaling factor to the given body segments in the model.");
+            constructProperty_bodies();
+            constructProperty_scale_factors(SimTK::Vec3{1.0});
+            constructProperty_preserve_masses(false);
+        }
+
+    private:
+        std::vector<ScalingStepValidationMessage> implValidate(
+            ScalingCache&,
+            const ScalingParameters& parameters,
+            const OpenSim::Model& sourceModel) const final
+        {
+            std::vector<ScalingStepValidationMessage> messages;
+
+            // Ensure every entry in `wrap_cylinders` can be found in the source model.
+            for (int i = 0; i < getProperty_bodies().size(); ++i) {
+                const auto* body = FindComponent<OpenSim::Body>(sourceModel, get_bodies(i));
+                if (not body) {
+                    std::stringstream msg;
+                    msg << get_bodies(i) << ": Cannot find a `Body` in 'bodies' in the source model (or it isn't a `Body`).";
+                    messages.emplace_back(ScalingStepValidationState::Error, std::move(msg).str());
+                }
+            }
+
+            const std::optional<double> blendingFactor = parameters.lookup<double>("blending_factor");
+            OSC_ASSERT_ALWAYS(blendingFactor && "blending_factor was not set by the warping engine");
+
+            return messages;
+        }
+
+        void implForEachScalingParameterDeclaration(const std::function<void(const ScalingParameterDeclaration&)>& callback) const final
+        {
+            callback(ScalingParameterDeclaration{"blending_factor", 1.0});
+        }
+
+        void implApplyScalingStep(
+            ScalingCache&,
+            const ScalingParameters& parameters,
+            const OpenSim::Model&,
+            OpenSim::Model& model) const final
+        {
+            const std::optional<double> blendingFactor = parameters.lookup<double>("blending_factor");
+            OSC_ASSERT_ALWAYS(blendingFactor && "blending_factor was not set by the warping engine");
+
+            OpenSim::ScaleSet scaleSet;
+            for (int i = 0; i < getProperty_bodies().size(); ++i) {
+                const auto* body = FindComponent<OpenSim::Body>(model, get_bodies(i));
+                if (not body) {
+                    std::stringstream msg;
+                    msg << get_bodies(i) << ": Cannot find a `Body` in 'bodies' in the source model (or it isn't a `Body`).";
+                    throw std::runtime_error{std::move(msg).str()};
+                }
+
+                OpenSim::Scale scale;
+                scale.setSegmentName(body->getName());
+                scale.setScaleFactors(get_scale_factors());
+                scale.setApply(true);
+                scaleSet.cloneAndAppend(scale);
+            }
+            model.scale(model.updWorkingState(), scaleSet, get_preserve_masses());
+            InitializeModel(model);
+            InitializeState(model);
+        }
+    };
+
     class RecalculateWrapCylinderRadiusFromStationScalingStep final : public ScalingStep {
         OpenSim_DECLARE_CONCRETE_OBJECT(RecalculateWrapCylinderRadiusFromStationScalingStep, ScalingStep)
     public:
@@ -1325,6 +1401,7 @@ Uses the Thin-Plate Spline (TPS) warping algorithm to scale `WrapCylinder`s in t
         ThinPlateSplineMeshSubstitutionScalingStep,
         ThinPlateSplineWrapCylinderScalingStep,
         ModelMassToSubjectMassScalingStep,
+        ManuallyScaleBodySegmentsScalingStep,
         RecalculateWrapCylinderRadiusFromStationScalingStep,
         RecalculateWrapCylinderXYZBodyRotationFromStationScalingStep
     >;